users/apiv1/serializers/artists_serializer.py
from rest_framework.serializers import ModelSerializer
from users.models import ArtistProfile

class ArtistProfileSerializer(ModelSerializer):
     class Meta:
         model = ArtistProfile
         fields = '__all__'