requirements.txt
asgiref==3.2.3
Babel==2.8.0
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0
django-allauth==0.41.0
django-appconf==1.0.3
django-axes==5.2.2
django-cors-headers==3.2.0
django-datetimepicker==3.14
django-ipware==2.1.0
django-phonenumber-field==4.0.0
django-templated-mail==1.1.1
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
idna==2.8
oauthlib==3.1.0
phonenumbers==8.11.1
Pillow==7.0.0
PyJWT==1.7.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.8
wincertstore==0.2
